CustomEntryTypesManager manages storing and loading custom entry types from the preferences and parsing/saving custom entry types from/to strings.
