+++ src/main/java/net/sf/jabref/bibtex/BibtexEntryWriter.java
@@ -30,6 +30,7 @@ import java.io.Writer;
 import java.util.*;
 
 import com.google.common.base.Strings;
+import net.sf.jabref.model.entry.EntryTypes;
 
 public class BibtexEntryWriter {
 
@@ -68,7 +69,7 @@ public class BibtexEntryWriter {
         // XXX JK: Look for all used field names not only defined once, since
         //         there may be some unofficial field name used.
         int max = 0;
-        for (BibtexEntryType type : BibtexEntryType.getAllValues()) {
+        for (BibtexEntryType type : EntryTypes.getAllValues()) {
             if (type.getRequiredFields() != null) {
                 for (String field : type.getRequiredFields()) {
                     max = Math.max(max, field.length());
