+++ src/main/java/net/sf/jabref/gui/EntryTypeList.java
@@ -28,6 +28,7 @@ import javax.swing.event.ListSelectionListener;
 import net.sf.jabref.model.entry.BibtexEntryType;
 import net.sf.jabref.model.entry.CustomEntryType;
 import net.sf.jabref.logic.l10n.Localization;
+import net.sf.jabref.model.entry.EntryTypes;
 import net.sf.jabref.util.Util;
 
 /**
@@ -92,7 +93,7 @@ public class EntryTypeList extends FieldSetComponent implements ListSelectionLis
         }
         for (int i = 0; i < selected.length; i++) {
             String typeName = listModel.get(selected[selected.length - 1 - i]);
-            BibtexEntryType type = BibtexEntryType.getType(typeName);
+            BibtexEntryType type = EntryTypes.getType(typeName);
 
             // If it is a custom entry type, we can remove it. If type == null, it means
             // the user must have added it and not yet applied it, so we can remove it
@@ -115,9 +116,9 @@ public class EntryTypeList extends FieldSetComponent implements ListSelectionLis
     public void enable(String typeName, boolean isChanged) {
         //String s = (String)list.getSelectedValue();
 
-        if (BibtexEntryType.getStandardType(typeName) != null) {
+        if (EntryTypes.getStandardType(typeName) != null) {
 
-            if (isChanged || (BibtexEntryType.getType(typeName) instanceof CustomEntryType)) {
+            if (isChanged || (EntryTypes.getType(typeName) instanceof CustomEntryType)) {
                 def.setEnabled(true);
             } else {
                 def.setEnabled(false);
