+++ src/main/java/net/sf/jabref/gui/labelPattern/LabelPatternPanel.java
@@ -33,7 +33,6 @@ import javax.swing.JScrollPane;
 import javax.swing.JTextField;
 
 import net.sf.jabref.gui.IconTheme;
-import net.sf.jabref.model.entry.BibtexEntryType;
 import net.sf.jabref.gui.GUIGlobals;
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
@@ -44,6 +43,7 @@ import net.sf.jabref.logic.labelPattern.GlobalLabelPattern;
 import net.sf.jabref.logic.util.strings.StringUtil;
 import net.sf.jabref.gui.help.HelpAction;
 import net.sf.jabref.gui.help.HelpDialog;
+import net.sf.jabref.model.entry.EntryTypes;
 
 public class LabelPatternPanel extends JPanel {
 
@@ -120,7 +120,7 @@ public class LabelPatternPanel extends JPanel {
         gbl.setConstraints(btnDefault, con);
         pan.add(btnDefault);
 
-        for (String s : BibtexEntryType.getAllTypes()) {
+        for (String s : EntryTypes.getAllTypes()) {
             textFields.put(s, addEntryType(pan, s, y));
             y++;
         }
