+++ src/main/java/net/sf/jabref/gui/preftabs/PreferencesDialog.java
@@ -47,8 +47,10 @@ import net.sf.jabref.gui.GUIGlobals;
 import net.sf.jabref.gui.MainTable;
 
 import com.jgoodies.forms.builder.ButtonBarBuilder;
+import net.sf.jabref.logic.CustomEntryTypesManager;
 import net.sf.jabref.logic.l10n.Localization;
-import net.sf.jabref.model.entry.BibtexEntryType;
+import net.sf.jabref.model.entry.CustomEntryType;
+import net.sf.jabref.model.entry.EntryTypes;
 import net.sf.jabref.util.Util;
 
 /**
@@ -213,7 +215,7 @@ public class PreferencesDialog extends JDialog {
                 try {
                     prefs.importPreferences(filename);
                     setValues();
-                    BibtexEntryType.loadCustomEntryTypes(prefs);
+                    CustomEntryTypesManager.loadCustomEntryTypes(prefs);
                     ExportFormats.initAllExports();
                     frame.removeCachedEntryEditors();
                     Globals.prefs.updateEntryEditorTabList();
