+++ src/main/java/net/sf/jabref/importer/CheckForNewEntryTypesAction.java
@@ -24,6 +24,7 @@ import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.model.entry.BibtexEntryType;
 import net.sf.jabref.model.entry.CustomEntryType;
 import net.sf.jabref.logic.l10n.Localization;
+import net.sf.jabref.model.entry.EntryTypes;
 
 /**
  * This action checks whether any new custom entry types were loaded from this
@@ -36,7 +37,7 @@ public class CheckForNewEntryTypesAction implements PostOpenAction {
         // See if any custom entry types were imported, but disregard those we already know:
         for (Iterator<String> i = pr.getEntryTypes().keySet().iterator(); i.hasNext();) {
             String typeName = i.next().toLowerCase();
-            if (BibtexEntryType.getType(typeName) != null) {
+            if (EntryTypes.getType(typeName) != null) {
                 i.remove();
             }
         }
@@ -63,7 +64,7 @@ public class CheckForNewEntryTypesAction implements PostOpenAction {
         if (answer == JOptionPane.YES_OPTION) {
             // Import
             for (BibtexEntryType typ : pr.getEntryTypes().values()) {
-                BibtexEntryType.addOrModifyCustomEntryType((CustomEntryType) typ);
+                EntryTypes.addOrModifyCustomEntryType((CustomEntryType) typ);
             }
         }
     }
