+++ src/main/java/net/sf/jabref/importer/fileformat/RepecNepImporter.java
@@ -26,9 +26,9 @@ import java.util.*;
 
 import net.sf.jabref.importer.ImportFormatReader;
 import net.sf.jabref.importer.OutputPrinter;
+import net.sf.jabref.model.entry.EntryTypes;
 import net.sf.jabref.model.entry.IdGenerator;
 import net.sf.jabref.model.entry.BibtexEntry;
-import net.sf.jabref.model.entry.BibtexEntryType;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
@@ -443,7 +443,7 @@ public class RepecNepImporter extends ImportFormat {
                 }
                 if (isStartOfWorkingPaper()) {
                     BibtexEntry be = new BibtexEntry(IdGenerator.next());
-                    be.setType(BibtexEntryType.getType("techreport"));
+                    be.setType(EntryTypes.getType("techreport"));
                     paperNoStr = this.lastLine.substring(0, this.lastLine.indexOf('.'));
                     parseTitleString(be);
                     if (startsWithKeyword(RepecNepImporter.recognizedFields)) {
