+++ src/main/java/net/sf/jabref/importer/fileformat/SixpackImporter.java
@@ -24,6 +24,7 @@ import java.util.HashMap;
 
 import net.sf.jabref.importer.ImportFormatReader;
 import net.sf.jabref.importer.OutputPrinter;
+import net.sf.jabref.model.entry.EntryTypes;
 import net.sf.jabref.model.entry.IdGenerator;
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.model.entry.BibtexEntryType;
@@ -136,8 +137,7 @@ public class SixpackImporter extends ImportFormat {
                 {
                     continue; // Avoid ArrayIndexOutOfBoundsException
                 }
-                BibtexEntryType typ = BibtexEntryType
-                        .getType(fields[1].toLowerCase());
+                BibtexEntryType typ = EntryTypes.getType(fields[1].toLowerCase());
                 if (typ == null) {
                     String type = "";
                     if (fields[1].equals("Masterthesis")) {
@@ -152,7 +152,7 @@ public class SixpackImporter extends ImportFormat {
                     if (fields[1].equals("Conference")) {
                         type = "proceedings";
                     }
-                    typ = BibtexEntryType.getType(type.toLowerCase());
+                    typ = EntryTypes.getType(type.toLowerCase());
                 }
                 entry = new BibtexEntry(IdGenerator.next(), typ);
                 String fld;
