+++ src/main/java/net/sf/jabref/logic/xmp/XMPSchemaBibtex.java
@@ -22,12 +22,8 @@ import javax.xml.transform.TransformerException;
 
 import net.sf.jabref.*;
 
-import net.sf.jabref.model.entry.AuthorList;
-import net.sf.jabref.model.entry.IdGenerator;
+import net.sf.jabref.model.entry.*;
 import net.sf.jabref.model.database.BibtexDatabase;
-import net.sf.jabref.model.entry.BibtexEntry;
-import net.sf.jabref.model.entry.BibtexEntryType;
-import net.sf.jabref.model.entry.BibtexEntryTypes;
 import org.apache.jempbox.xmp.XMPMetadata;
 import org.apache.jempbox.xmp.XMPSchema;
 import org.w3c.dom.Element;
@@ -306,7 +302,7 @@ public class XMPSchemaBibtex extends XMPSchema {
         String type = getTextProperty("entrytype");
         BibtexEntryType t;
         if (type != null) {
-            t = BibtexEntryType.getStandardType(type);
+            t = EntryTypes.getStandardType(type);
         } else {
             t = BibtexEntryTypes.OTHER;
         }
