+++ src/main/java/net/sf/jabref/model/entry/BibtexEntryTypes.java
@@ -10,7 +10,7 @@ public class BibtexEntryTypes {
     // Get an entry type defined in BibtexEntryType
     public static BibtexEntryType getEntryType(String type) {
         // decide which entryType object to return
-        Object o = BibtexEntryType.getType(type);
+        Object o = EntryTypes.getType(type);
         if (o != null) {
             return (BibtexEntryType) o;
         } else {
